server/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
PyMySQL==1.1.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.5
python-multipart==0.0.21
pydantic-settings==2.12.0
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1
python-dotenv==1.2.1